+++ Core/Models/Usuario.cs
@@ -6,9 +6,9 @@ namespace Core.Models
     {
         public int Id { get; set; }
         public string Nome { get; set; }
-        public int Cpf { get; set; }
+        public long Cpf { get; set; }
         public DateTime DataNascimento { get; set; }
-        public int Telefone { get; set; }
+        public long Telefone { get; set; }
         public string Email { get; set; }
         public int Cep { get; set; }
         public string Senha { get; set; }
